Unit test for formatDate #77
