--- __tests__/example.test.js
@@ -1,3 +1,0 @@
-test("basic math test", () => {
-  expect(1 + 2).toBe(3);
-});
