+++ js/Analytics/formatDate.js
@@ -1,0 +1,11 @@
+function formatDate(date) {
+  if (!date) return "No date";
+  return date.toLocaleDateString("en-US", {
+    year: "numeric",
+    month: "short",
+    day: "numeric",
+    hour: "2-digit",
+    minute: "2-digit",
+  });
+}
+module.exports = formatDate;
